point组件抽离
